Convert ScreenshotHolder to an interface for testing purpose.
