+++ features/rageshake/src/main/kotlin/io/element/android/features/rageshake/screenshot/ScreenshotHolder.kt
@@ -1,5 +1,5 @@
 /*
- * Copyright (c) 2022 New Vector Ltd
+ * Copyright (c) 2023 New Vector Ltd
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -16,29 +16,11 @@
 
 package io.element.android.features.rageshake.screenshot
 
-import android.content.Context
 import android.graphics.Bitmap
-import io.element.android.libraries.androidutils.bitmap.writeBitmap
-import io.element.android.libraries.androidutils.file.safeDelete
-import io.element.android.libraries.di.AppScope
-import io.element.android.libraries.di.ApplicationContext
-import io.element.android.libraries.di.SingleIn
 import java.io.File
-import javax.inject.Inject
 
-@SingleIn(AppScope::class)
-class ScreenshotHolder @Inject constructor(
-    @ApplicationContext private val context: Context,
-) {
-    private val file = File(context.filesDir, "screenshot.png")
-
-    fun writeBitmap(data: Bitmap) {
-        file.writeBitmap(data, Bitmap.CompressFormat.PNG, 85)
-    }
-
-    fun getFile() = file.takeIf { it.exists() && it.length() > 0 }
-
-    fun reset() {
-        file.safeDelete()
-    }
+interface ScreenshotHolder {
+    fun writeBitmap(data: Bitmap)
+    fun getFile(): File?
+    fun reset()
 }
